+++ src/main/kotlin/kotools/assert/Annotations.kt
@@ -1,5 +1,8 @@
 package kotools.assert
 
+import org.junit.jupiter.api.Nested
+import org.junit.jupiter.api.Test
+
 /**
  * Specifies the first [version] of Kotools Assert where a declaration has
  * appeared.
@@ -21,3 +24,9 @@ package kotools.assert
     AnnotationTarget.TYPEALIAS
 )
 internal annotation class SinceKotoolsAssert(val version: String)
+
+@SinceKotoolsAssert("2.1")
+public typealias Nested = Nested
+
+@SinceKotoolsAssert("2.1")
+public typealias Test = Test